Create name_preference
Liquid for customers working with people whose preferred name may differ from their legal name. The visitor is redirected to a page (in this instance, after a donation) that tells them the name on file for them and includes a link to their public profile to edit their name if they'd like.