First draft
This liquid will run through pages with a certain tag, then display a preview in another page. There are two steps: Create the snipped and save it under "files" of a custom theme, and drop in the liquid to whichever page you would like the list of tagged pages to appear.